Switch `.vimrc` to lua
This is a sad day and will likely mean a thousand tiny edits in the
coming weeks to fix things the way I like them. But as I'm writing more
ypescript and playing with integrations of local LLMs for code
completion, it's looking like a good time to switch out from a vim
config that's written in vimscript compared to lua.

Do I regret my decision? Yes. When I started off my career in coding I
knowingly front-loaded a lot of time learning and setting up a vim
config that I knew and loved so that I could have slightly better work
habits and tools for years to come.

Having to go over it again (at least in an only slightly diluted form!)
is making me rethink this ROI.